+++ src/starkbank/event.clj
@@ -69,7 +69,7 @@
     - `:after` [string, default nil]: date filter for maps created only after specified date. ex: ~D[2020-03-25]
     - `:before` [string, default nil]: date filter for maps created only before specified date. ex: ~D[2020-03-25]
     - `:is-delivered` [bool, default nil]: filter successfully delivered events. ex: true or false
-    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.user/set has not been set.
+    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.settings/set-default-user has not been set.
 
   ## Return:
     - stream of Event maps with updated attributes"
@@ -91,7 +91,7 @@
     - `id` [string]: map unique id. ex: \"5656565656565656\"
 
   ## Options:
-    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.user/set has not been set.
+    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.settings/set-default-user has not been set.
 
   ## Return:
     - Event map with updated attributes"
@@ -112,7 +112,7 @@
     - `id` [string]: Event unique id. ex: \"5656565656565656\"
 
   ## Options:
-    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.user/set has not been set.
+    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.settings/set-default-user has not been set.
 
   ## Return:
     - deleted Event map with updated attributes"
@@ -135,7 +135,7 @@
     - `:is-delivered` [bool]: If true and event hasn't been delivered already, event will be set as delivered. ex: true
 
   ## Parameters (optional):
-    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.user/set has not been set.
+    - `:user` [Project]: Project map returned from starkbank.user/project. Only necessary if starkbank.settings/set-default-user has not been set.
 
   ## Return:
     - target Event with updated attributes"